requirements.txt
# API Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Document Processing
pymupdf==1.24.3
pdfplumber==0.11.0
pillow==10.3.0
pytesseract==0.3.10

# Email Parsing
mail-parser==3.15.0

# RAG / Vector Store
langchain==0.2.1
langchain-community==0.2.1
langchain-openai==0.1.7
langchain-chroma==0.1.1
chromadb==0.5.0

# Embeddings & LLM
openai==1.30.1
tiktoken==0.7.0

# Database (structured metadata)
sqlalchemy==2.0.30
alembic==1.13.1

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Utilities
python-dotenv==1.0.1
python-dateutil==2.9.0
tenacity==8.3.0
httpx==0.27.0
aiofiles==23.2.1
tqdm==4.66.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
